feat: export budget in pdf report

+++ src/main/java/org/sistema/arroz/riceservice/modules/supplyFormula/application/port/in/GetSuppliesFormulasDTO.java
@@ -10,6 +10,6 @@ public class GetSuppliesFormulasDTO {
     Long supplyId;
     Long productId;
     String supplyName;
-    Integer cantForHectare;
+    Double cantForHectare;
     SupplyMetricType supplyMetricType;
 }
